requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.62.0
geopandas>=0.10.0
shapely>=1.8.0
pyarrow>=10.0.0
wandb>=0.13.0